pyproject.toml
[project]
name = "agencia-v3"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "ollama>=0.2.0",
  "groq>=0.9.0",
  "google-genai>=1.0.0",
  "mistralai>=1.0.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "python-dotenv>=1.0.0",
  "aiosqlite>=0.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-httpx>=0.30.0",
]
